split config group to files
